requirements.txt
git+https://github.com/Rapptz/discord.py
Pillow
asyncio
aiomysql
psutil
requests